requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
html5lib==1.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
